requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
sse_starlette
pypdf
chromadb
faiss-cpu
langchainhub
PyPDF2
langchain-objectbox
langchain
langchain-pinecone
ragas
langchain-google-genai
langchain-google-vertexai
transformers
langchain_huggingface
langchain-groq